cpufreq: cafactive: apply hispeed boost only on cpu0 and cpu4
Boost in interactive is intended to ramp up the frequency on sudden
workload burst, and since most *sudden workload* is single-threaded,
boosting up exclusively on cpu0 is quite enough.

Rest of the cores can slowly ramp up and match the actual workload.
This is not recommended on devices where reducing latency is very
important.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: Tkkg1994 <luca.grifo@outlook.com>
Signed-off-by: djb77 <dwayne.bakewell@gmail.com>
Signed-off-by: morogoku <morogoku@hotmail.com>